Feat/21 crud (#28)
* feat: 공구방 상세페이지 CRUD 구현

더 상세한 제어 조건은 추후 추가

* feat: 상세페이지 공구방삭제 및 입금폼작성

공구방삭제 프론트단 구현
입금폼제출 라우팅

* feat: 새소식 리스팅 및 새소식 글쓰기 구현

* feat: 실시간 채팅 구현

DB CRUD는 아직 진행하지 않음

* feat: 일부 수정사항

* feat: 루트 디렉토리에 .gitignore 추가

* feat: SQL 외래키 오류 해결

+++ server/models/cobuying_room.js
@@ -13,13 +13,7 @@ module.exports = (sequelize, Sequelize) => {
         allowNull: false,
       },
       state: {
-        type: Sequelize.ENUM(
-          "demand",
-          "deposit",
-          "ready",
-          "delivery",
-          "complete"
-        ),
+        type: Sequelize.ENUM("demand", "deposit", "ready", "delivery", "complete"),
         defaultValue: "demand",
         comment: "수요조사, 입금중, 준비중, 배송중, 공동구매 완료만 허용",
       },
@@ -28,7 +22,7 @@ module.exports = (sequelize, Sequelize) => {
         allowNull: false,
       },
       host_id: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.BIGINT,
       },
     },
     {
